lib/agents/multi-tenant-architecture.ts
import { BaseAgent, AgentConfig } from './base';

export interface Multi-tenantArchitectureConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class Multi-tenantArchitectureAgent extends BaseAgent {
  private config: Multi-tenantArchitectureConfig;

  constructor(config: Multi-tenantArchitectureConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Multi-tenant Architecture implementation
    await new Promise(resolve => setTimeout(resolve, 1000));
    
    return `Multi-tenant Architecture completed successfully for: ${input}`;
  }

  async process(data: any): Promise<any> {
    // Advanced multi-tenant architecture processing
    return { status: 'success', data };
  }
}

// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-01-08
import { BaseAgent, AgentConfig } from './base';

export interface Multi-tenantArchitectureConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class Multi-tenantArchitectureAgent extends BaseAgent {
  private config: Multi-tenantArchitectureConfig;

  constructor(config: Multi-tenantArchitectureConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Multi-tenant Architecture impleme...


// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-01-07
import { BaseAgent, AgentConfig } from './base';

export interface Multi-tenantArchitectureConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class Multi-tenantArchitectureAgent extends BaseAgent {
  private config: Multi-tenantArchitectureConfig;

  constructor(config: Multi-tenantArchitectureConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Multi-tenant Architecture impleme...


// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-02-17
import { BaseAgent, AgentConfig } from './base';

export interface Multi-tenantArchitectureConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class Multi-tenantArchitectureAgent extends BaseAgent {
  private config: Multi-tenantArchitectureConfig;

  constructor(config: Multi-tenantArchitectureConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Multi-tenant Architecture impleme...


// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-04-22
import { BaseAgent, AgentConfig } from './base';

export interface Multi-tenantArchitectureConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class Multi-tenantArchitectureAgent extends BaseAgent {
  private config: Multi-tenantArchitectureConfig;

  constructor(config: Multi-tenantArchitectureConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Multi-tenant Architecture impleme...
