__tests__/lib/agents/multi-tenant-architecture.test.ts
import { Multi-tenantArchitectureAgent } from '@/lib/agents/multi-tenant-architecture';

describe('Multi-tenant ArchitectureAgent', () => {
  let agent: Multi-tenantArchitectureAgent;

  beforeEach(() => {
    agent = new Multi-tenantArchitectureAgent({
      name: 'Test Multi-tenant Architecture Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      const result = await agent.execute('test input');
      expect(result).toContain('Multi-tenant Architecture completed successfully');
    });

    it('should handle empty input', async () => {
      const result = await agent.execute('');
      expect(result).toBeDefined();
    });
  });

  describe('process', () => {
    it('should process data correctly', async () => {
      const result = await agent.process({ test: 'data' });
      expect(result.status).toBe('success');
    });
  });
});

// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-01-08
import { Multi-tenantArchitectureAgent } from '@/lib/agents/multi-tenant-architecture';

describe('Multi-tenant ArchitectureAgent', () => {
  let agent: Multi-tenantArchitectureAgent;

  beforeEach(() => {
    agent = new Multi-tenantArchitectureAgent({
      name: 'Test Multi-tenant Architecture Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input succe...


// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-01-07
import { Multi-tenantArchitectureAgent } from '@/lib/agents/multi-tenant-architecture';

describe('Multi-tenant ArchitectureAgent', () => {
  let agent: Multi-tenantArchitectureAgent;

  beforeEach(() => {
    agent = new Multi-tenantArchitectureAgent({
      name: 'Test Multi-tenant Architecture Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input succe...


// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-02-17
import { Multi-tenantArchitectureAgent } from '@/lib/agents/multi-tenant-architecture';

describe('Multi-tenant ArchitectureAgent', () => {
  let agent: Multi-tenantArchitectureAgent;

  beforeEach(() => {
    agent = new Multi-tenantArchitectureAgent({
      name: 'Test Multi-tenant Architecture Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input succe...


// FEATURE UPDATE: Multi-tenant Architecture
// Date: 2024-04-22
import { Multi-tenantArchitectureAgent } from '@/lib/agents/multi-tenant-architecture';

describe('Multi-tenant ArchitectureAgent', () => {
  let agent: Multi-tenantArchitectureAgent;

  beforeEach(() => {
    agent = new Multi-tenantArchitectureAgent({
      name: 'Test Multi-tenant Architecture Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input succe...
